ekranornek/CARPETOS_V2/WEB_APP/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-socketio==5.10.0
eventlet==0.33.3
Werkzeug==3.0.1

